requirements.txt
attrs==23.2.0
Automat==22.10.0
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==42.0.8
cssselect==1.2.0
defusedxml==0.7.1
filelock==3.15.1
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
itemadapter==0.9.0
itemloaders==1.3.1
jmespath==1.0.1
json5==0.9.25
lxml==5.2.2
packaging==24.1
parsel==1.9.1
Protego==0.3.1
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyDispatcher==2.0.7
PyMySQL==1.1.1
pyOpenSSL==24.1.0
python-dotenv==1.0.1
queuelib==1.7.0
requests==2.32.3
requests-file==2.1.0
scrapeops_scrapy==0.5.4
scrapeops_scrapy_proxy_sdk==1.0
Scrapy==2.11.2
scrapy-user-agents==0.1.1
service-identity==24.1.0
six==1.16.0
tld==0.13
tldextract==5.1.2
Twisted==24.3.0
typing_extensions==4.12.2
ua-parser==0.18.0
urllib3==2.2.2
user-agents==2.2.0
w3lib==2.2.1
zope.interface==6.4.post2
